fix/refactorAll
