+++ src/components/Artwork/Card.js
@@ -15,9 +15,6 @@ function Card({
 }) {
   const navigate = useNavigate();
 
-  //이제 수정 해야함
-  //왜 저장이 안되지..
-  // ㅇㅁㄹㅇ
   return (
     <div
       className="card"
@@ -32,7 +29,6 @@ function Card({
           backgroundSize: 'cover',
         }}
       />
-      {/* <img src={img_url} className="cardImg" /> */}
       <div className="cardTitle">
         <span className="cardTitleSpan">{title}</span>
         <div className="cardProfileDiv">
