--- src/components/Artwork/PeopleSearch.js
@@ -1,7 +1,0 @@
-import React from 'react';
-
-function PeopleSearch() {
-  return <></>;
-}
-
-export default PeopleSearch;
