+++ src/components/Upload/UploadWrite.js
@@ -12,11 +12,9 @@ function App({
   startFetch,
   previewImg,
 }) {
-  const [tagModalState, setTagModalState] = useState(false); // 태그 추가창 열고 닫기
-  const [uploadModalState, setuploadModalState] = useState(false); // ㄹ
+  const [tagModalState, setTagModalState] = useState(false);
+  const [uploadModalState, setuploadModalState] = useState(false);
   const [uploadTagData, setUploadTagData] = useState([]);
-
-  const [previewState, setPreviewState] = useState(false);
 
   const titleValue = useRef();
   const contentValue = useRef();
@@ -31,7 +29,6 @@ function App({
     const updateTagMockData = [];
     if (tagFieldRef.current) {
       tagFieldRef.current.childNodes.forEach(element => {
-        // 고차함수 사용이 안돼서 forEach를 사용했던 것 같다
         updateTagMockData.push(element.innerText);
       });
     }
@@ -39,8 +36,7 @@ function App({
   }
 
   function addTag(event) {
-    if (tagModalState) setTagModalState(false);
-    if (!tagModalState) setTagModalState(true);
+    setTagModalState(!tagModalState);
   }
 
   function clickPreveiw() {
