+++ src/components/Upload/categoryModal.scss
