+++ src/components/Upload/uploadMain.scss
